Update Default Content Docs Link To Link to Docs
Updates the Read the docs URL from directing to the homepage of
tailwindcss.com to the documentation route within the tailwindcss.com
site.

The link directing to the homepage, rather than the documentation page
feels like a misalignment of what the link describes, which lead to the
decision to change the URL of the link to direct to the documentation
page.